+++ src/config.ts
@@ -21,12 +21,12 @@ import { createAPIClient } from './client';
  * `instance.config` in a UI.
  */
 export const instanceConfigFields: IntegrationInstanceConfigFieldMap = {
-  clientId: {
-    type: 'string',
-  },
-  clientSecret: {
+  apiToken: {
     type: 'string',
     mask: true,
+  },
+  orgName: {
+    type: 'string',
   },
 };
 
@@ -36,14 +36,13 @@ export const instanceConfigFields: IntegrationInstanceConfigFieldMap = {
  */
 export interface IntegrationConfig extends IntegrationInstanceConfig {
   /**
-   * The provider API client ID used to authenticate requests.
+   * The provider access token used to authenticate requests.
    */
-  clientId: string;
-
+  apiToken: string;
   /**
-   * The provider API client secret used to authenticate requests.
+   * The org name used for the hostname.
    */
-  clientSecret: string;
+  orgName: string;
 }
 
 export async function validateInvocation(
@@ -51,9 +50,9 @@ export async function validateInvocation(
 ) {
   const { config } = context.instance;
 
-  if (!config.clientId || !config.clientSecret) {
+  if (!config.apiToken || !config.orgName) {
     throw new IntegrationValidationError(
-      'Config requires all of {clientId, clientSecret}',
+      'Config requires all of { apiToken, orgName }',
     );
   }